big chances with little effect
now the data package saves the untranslated coordinates and widths and
uses a coordinate translator from the adaptar to calculate the
coordinates for the gui.

+++ src/org/jdesktop/wonderland/modules/oweditor/client/editor/data/AdapterObserver.java
@@ -38,7 +38,7 @@ public class AdapterObserver implements AdapterObserverInterface{
         dm.createNewObject(dataObject);
     }
 
-    public void notifyTranslation(long id, int x, int y, int z) {
+    public void notifyTranslation(long id, float x, float y, float z) {
         dm.updateTranslation(id, x, y, z);
     }
