+++ src/org/jdesktop/wonderland/modules/oweditor/client/editor/data/Vector3D.java
@@ -8,9 +8,9 @@ package org.jdesktop.wonderland.modules.oweditor.client.editor.data;
  */
 public class Vector3D {
     
-    protected int x = 0;
-    protected int y = 0;
-    protected int z = 0;
+    protected float x = 0;
+    protected float y = 0;
+    protected float z = 0;
     
     /**
      * Creates a new vector instance.
@@ -19,7 +19,7 @@ public class Vector3D {
      * @param y the y coordinate.
      * @param z the z coordinate.
      */
-    public Vector3D(int x, int y, int z){
+    public Vector3D(float x, float y, float z){
         this.x = x;
         this.y = y;
         this.z = z;
